fix: Refactor project folder structure
