+++ lib/src/routes/routes.dart
@@ -1,11 +1,9 @@
-import 'package:delivery_app/src/pages/forgot_password_page.dart';
-import 'package:delivery_app/src/pages/login_page.dart';
-import 'package:delivery_app/src/pages/sign_up_page.dart';
-import 'package:delivery_app/src/tabs/tabs_page.dart';
+import 'package:delivery_app/src/features/presentation/forgot_password_page/view/forgot_password_page.dart';
+import 'package:delivery_app/src/features/presentation/login_page/view/login_page.dart';
+import 'package:delivery_app/src/features/presentation/sign_up_page/view/sign_up_page.dart';
+import 'package:delivery_app/src/features/presentation/tabs/tabs_page.dart';
+import 'package:delivery_app/src/features/presentation/welcome_page/view/welcome_page.dart';
 import 'package:flutter/widgets.dart';
-
-// Pages
-import 'package:delivery_app/src/pages/welcome_page.dart';
 
 final routes = <String, WidgetBuilder>{
   'welcome': (BuildContext context) => WelcomePage(),
